Eric Sunshine eliminated several compilation warnings from DirectSound 3D renderer.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@14592 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/sound/renderer/ds3d/sndhdl.cpp
@@ -30,7 +30,7 @@ csSoundHandleDS3D::csSoundHandleDS3D(csRef<csSoundRenderDS3D> srdr, csRef<iSound
   buffer=0;  
 
   NumSamples = Data->IsStatic() ? Data->GetStaticSampleCount() :
-  (((float)(Data->GetFormat()->Freq)) * BufferLengthSeconds);
+  (long)(Data->GetFormat()->Freq * BufferLengthSeconds);
 
   buffer_length=(NumSamples* Data->GetFormat()->Bits * Data->GetFormat()->Channels)/8;
   if (LocalBuffer)
